Mutate existing default export declarations for improved compatibility
Calling `t.exportDefaultDeclaration` triggers the `ExportDefaultDeclaration` in other babel transforms. In some cases triggering that hook multiple times can lead to unexpected behavior.

For example, in the the template-colocation-plugin at https://github.com/embroider-build/embroider/blob/289f83d80d/packages/shared-internals/src/template-colocation-plugin.ts), the first call of the hook will see a class declaration and set `state.associate`[^1] (which later causes `setComponentTemplate` to be appended after the `export` [^2]). The second call of the hook sees a non-class expression and wraps it in `setComponentTemplate` within the export [^3]. In the end, that leaves us with two `setComponentTemplate` calls.

[^1]: https://github.com/embroider-build/embroider/blob/289f83d80d/packages/shared-internals/src/template-colocation-plugin.ts#L103-L111
[^2]: https://github.com/embroider-build/embroider/blob/289f83d80d/packages/shared-internals/src/template-colocation-plugin.ts#L82-L91
[^3]: https://github.com/embroider-build/embroider/blob/289f83d80d/packages/shared-internals/src/template-colocation-plugin.ts#L129-L131